requirements.txt
flask
pymupdf
groq
python-dotenv
scikit-learn
numpy